hot_dogg/requirements.txt
numpy>=1.24
pandas>=2.0
python-dotenv>=1.0
websockets>=13.0
deriv-api>=1.0

